# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: B Y E C O M  is your ecom. shop solution, where you can sell anything
    online with ease and at an affordable price. We provide an eCommerce platform
    for businesses to sell their products online and connect with customers worldwide.
  title: B Y E C O M  LTD
  version: "1.0"
paths:
...
  /user/cart/checkout:
    post:
      description: Buying cart products
      parameters:
      - description: address id
        in: formData
        name: address
        required: true
        type: string
      - description: coupon code
        in: formData
        name: coupon
        type: string
      - description: payment method
        in: formData
        name: method
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Cart Chekout Only
      tags:
      - User Order
...
  /user/order:
    get:
      description: Showing Orders with its details
      produces:
      - application/json
      responses: {}
      summary: Show Orders
      tags:
      - User Order
    patch:
      description: Cancelling ordered products individually
      parameters:
      - description: order id
        in: query
        name: order
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Cancelling Order
      tags:
      - User Order
...
swagger: "2.0"